requirements.txt
cffi==1.14.6
cryptography==3.4.8
greenlet==1.1.1
paho-mqtt==1.5.1
passlib==1.7.4
pycparser==2.20
PyMySQL==1.0.2
six==1.16.0
SQLAlchemy==1.4.23
SQLAlchemy-Utils==0.37.8
